Move the theme config into App component
so that it won't be exported somewhere else
